feat: review rework
