+++ tests/ttnn/unit_tests/operations/eltwise/test_binaryng_ND.py
@@ -98,14 +98,12 @@ def test_ND_subtile_bcast(device, shapes, ttnn_fn):
                 input_tensor_a, input_tensor_b, memory_config=ttnn.DRAM_MEMORY_CONFIG, use_legacy=None
             )
 
-    output_tensor = ttnn.to_torch(output_tensor)
-
-    # Use ULP comparison for hypot, PCC for other operations
     if ttnn_fn == ttnn.hypot:
         from tests.ttnn.utils_for_testing import assert_with_ulp
 
-        assert_with_ulp(ttnn.from_torch(output_tensor, layout=ttnn.TILE_LAYOUT, device=device), torch_output_tensor)
+        assert_with_ulp(output_tensor, torch_output_tensor)
     else:
+        output_tensor = ttnn.to_torch(output_tensor)
         assert ttnn.pearson_correlation_coefficient(torch_output_tensor, output_tensor) >= 0.999
