+++ ttnn/cpp/ttnn/operations/eltwise/binary/binary.hpp
@@ -8,6 +8,7 @@
 #include "ttnn/decorators.hpp"
 #include "ttnn/operations/eltwise/unary/common/unary_op_types.hpp"
 #include "ttnn/operations/eltwise/binary/common/binary_op_types.hpp"
+#include "ttnn/operations/eltwise/unary/unary.hpp"
 
 namespace ttnn {
 namespace operations::binary {
@@ -161,6 +162,44 @@ struct BinaryOperationSubalpha {
         const std::optional<Tensor>& output = std::nullopt);
 };
 
+/**
+ * @brief Performs element-wise hypot operation: sqrt(a^2 + b^2).
+ * When inputs are Tensors, the supported dtypes are float32 and bfloat16.
+ *
+ * @param input_a The first input tensor
+ * @param input_b The second input tensor
+ * @return The result tensor
+ */
+struct ExecuteHypot {
+    // Main tensor-tensor interface
+    static Tensor invoke(
+        const Tensor& input_tensor_a,
+        const Tensor& input_tensor_b,
+        const std::optional<const DataType>& dtype = std::nullopt,
+        const std::optional<MemoryConfig>& memory_config = std::nullopt,
+        const std::optional<Tensor>& optional_output_tensor = std::nullopt,
+        tt::stl::Span<const unary::EltwiseUnaryWithParam> post_activations = {},
+        tt::stl::Span<const unary::EltwiseUnaryWithParam> lhs_activations = {},
+        tt::stl::Span<const unary::EltwiseUnaryWithParam> rhs_activations = {},
+        std::optional<bool> use_legacy = std::nullopt);
+
+    // Simple tensor-tensor interface
+    static Tensor invoke(
+        const Tensor& input_tensor_a, const Tensor& input_tensor_b, const std::optional<MemoryConfig>& memory_config);
+
+    // Tensor-scalar interfaces for pybind
+    static Tensor invoke(
+        const Tensor& input_tensor_a,
+        float input_tensor_b,
+        const std::optional<const DataType>& dtype = std::nullopt,
+        const std::optional<MemoryConfig>& memory_config = std::nullopt,
+        const std::optional<Tensor>& optional_output_tensor = std::nullopt,
+        tt::stl::Span<const unary::EltwiseUnaryWithParam> post_activations = {},
+        tt::stl::Span<const unary::EltwiseUnaryWithParam> lhs_activations = {},
+        tt::stl::Span<const unary::EltwiseUnaryWithParam> rhs_activations = {},
+        std::optional<bool> use_legacy = std::nullopt);
+};
+
 }  // namespace operations::binary
 
 constexpr auto add =
@@ -271,6 +310,7 @@ constexpr auto logical_right_shift = ttnn::register_operation<
     operations::binary::BinaryOperation<operations::binary::BinaryOpType::LOGICAL_RIGHT_SHIFT>>();
 constexpr auto xlogy = ttnn::
     register_operation<"ttnn::xlogy", operations::binary::BinaryOperation<operations::binary::BinaryOpType::XLOGY>>();
+constexpr auto hypot = ttnn::register_operation<"ttnn::hypot", operations::binary::ExecuteHypot>();
 
 template <typename InputBType>
 ttnn::Tensor operator+(const ttnn::Tensor& lhs, InputBType rhs) {
