+++ ttnn/cpp/ttnn/operations/eltwise/binary_ng/device/binary_ng_device_operation.cpp
@@ -51,7 +51,8 @@ bool is_binary_sfpu_op(BinaryOpType val, DataType a, DataType b) {
         case MAXIMUM:
         case MINIMUM:
         case XLOGY:
-        case POWER: return true;
+        case POWER:
+        case HYPOT: return true;
         default: return false;
     }
     return false;
